LayoutTests/fullscreen/full-screen-enabled-prefixed.html

<body>
<div>This tests the <code>fullscreenEnabled</code> property laid out in section 4 of the W3C 
<a href="http://dvcs.w3.org/hg/fullscreen/raw-file/tip/Overview.html">Full Screen API</a></div>
<script src="full-screen-test.js"></script>
<script>
    var iframe = document.documentElement.appendChild(document.createElement('iframe'));
    iframe.setAttribute('webkitallowfullscreen', 'true');
    var iframe2 = document.documentElement.appendChild(document.createElement('iframe'));
    testExpected('iframe.contentDocument.webkitFullscreenEnabled', true);
    testExpected('iframe2.contentDocument.webkitFullscreenEnabled', false);
    endTest();
</script>
